refactor(example): rename and restructure user cache related classes
- Rename UserCacheProxy to UserCache
- Rename CacheConfiguration to ClassDefinedCacheConfiguration
- Update related imports and usages across multiple files
- Simplify UserCache interface to extend Cache instead of ComputedCache
