+++ cocache-example/src/main/kotlin/me/ahoo/cache/example/AppServer.kt
@@ -12,7 +12,7 @@
  */
 package me.ahoo.cache.example
 
-import me.ahoo.cache.example.cache.UserCacheProxy
+import me.ahoo.cache.example.cache.UserCache
 import me.ahoo.cache.spring.EnableCoCache
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
@@ -22,7 +22,7 @@ import org.springframework.boot.runApplication
  *
  * @author ahoo wang
  */
-@EnableCoCache(caches = [UserCacheProxy::class])
+@EnableCoCache(caches = [UserCache::class])
 @SpringBootApplication
 class AppServer
