+++ cocache-example/src/main/kotlin/me/ahoo/cache/example/controller/TestController.kt
@@ -13,7 +13,7 @@
 package me.ahoo.cache.example.controller
 
 import me.ahoo.cache.CoherentCache
-import me.ahoo.cache.example.cache.UserCacheProxy
+import me.ahoo.cache.example.cache.UserCache
 import me.ahoo.cache.example.model.User
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.web.bind.annotation.GetMapping
@@ -21,7 +21,6 @@ import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import java.util.PrimitiveIterator
 import java.util.UUID
 
 /**
@@ -34,18 +33,18 @@ import java.util.UUID
 class TestController(
     @param:Qualifier("userCache")
     private val userCaching: CoherentCache<String, User>,
-    private val userCacheProxy: UserCacheProxy
+    private val userCache: UserCache
 ) {
 
     @GetMapping("{id}")
     fun get(@PathVariable id: String): User? {
-        return userCacheProxy[id]
+        return userCache[id]
     }
 
     @PostMapping("{id}")
     fun set(@PathVariable id: String): String {
         val user = User(id, UUID.randomUUID().toString())
-        userCacheProxy[user.id] = user
+        userCache[user.id] = user
         return user.id
     }
 }
